images

+++ android.py
@@ -1,40 +1,42 @@
-from datetime import datetime
 import os
 from os.path import basename
 import shutil
-from flask import Blueprint, current_app,  redirect, request, url_for
-from helpers import get_secure_filename_filepath
+from flask import Blueprint, request, current_app
+from flask.helpers import url_for
+from helpers import get_secure_filename_filepath, download_from_s3
+from werkzeug.utils import redirect
 from PIL import Image
 from zipfile import ZipFile
 
-bp = Blueprint('android', __name__, url_prefix='/android')
+bp = Blueprint('android', __name__)
 
-ICON_SIZES = [29, 40, 57, 60, 80, 87, 114, 120, 180, 1024]
-
-@bp.route('/', methods=["POST"])
+@bp.route('/android', methods=["POST"])
 def create_images():
-    filename = request.json['filename']
-    filename, filepath = get_secure_filename_filepath(filename)
+    if request.method == 'POST':
+        ICON_SIZES = [29, 40, 57, 58, 60, 80, 87, 114, 120, 180, 1024]
 
-    tempfolder = os.path.join(current_app.config['UPLOAD_FOLDER'], 'temp')
-    os.makedirs(tempfolder)
+        filename = request.json['filename']
+        
+        # filename, filepath = get_secure_filename_filepath(filename)
+        tempfolder = os.path.join(current_app.config['DOWNLOAD_FOLDER'], 'temp')
+        
+        if not os.path.exists(tempfolder):
+            os.makedirs(tempfolder)
 
-    for size in ICON_SIZES:
-       outfile = os.path.join(tempfolder, f'{size}.png')
-       image = Image.open(filepath)
-       out = image.resize((size, size))
-       out.save(outfile, 'PNG')
+        for size in ICON_SIZES:
+            file_stream = download_from_s3(filename)
+            image = Image.open(file_stream)
+            out = image.resize((size, size))
+            outfile = os.path.join(tempfolder, f'{size}.png')
+            out.save(outfile, "PNG")
 
-    now = datetime.now()
-    timestamp = str(datetime.timestamp(now)).rsplit('.')[0]
-    Zipfilename = f'{timestamp}.zip'
-    Zipfilepath = os.path.join(current_app.config['UPLOAD_FOLDER'], Zipfilename)
-
-    with ZipFile(Zipfilepath, 'w') as zipObj:
-        for foldername, subfolders, filenames, in os.walk(tempfolder):
-            for filepath in filenames:
-                filepath = os.path.join(foldername, filename)
-                zipObj.write(filepath, basename(filepath))
-        shutil.rmtree(tempfolder)
-        return redirect(url_for('download_file', name=Zipfilename ))
-
+        zipfilename = 'Icons.zip'
+        zipfilepath = os.path.join(current_app.config['DOWNLOAD_FOLDER'], zipfilename)
+        with ZipFile(zipfilepath, 'w') as zipObj:
+            for foldername, subfolders, filenames in os.walk(tempfolder):
+                for filename in filenames:
+                    filepath = os.path.join(foldername, filename)
+                    zipObj.write(filepath, basename(filepath))
+            shutil.rmtree(tempfolder)
+            print(zipObj.filename)
+            return redirect(url_for('download_file', name=zipfilename))
